feat: ajuste de pasta.
