--- app/(app)/dashboard.tsx
@@ -1,2 +1,0 @@
-import DashboardScreen from "../../src/features/dashboard/DashboardScreen";
-export default DashboardScreen;
