--- app/(app)/meus-cartoes.tsx
@@ -1,2 +1,0 @@
-import CardsScreen from "../../src/features/cards/CardsScreen";
-export default CardsScreen;
