--- app/(app)/outros-servicos.tsx
@@ -1,2 +1,0 @@
-import ServicesScreen from "../../src/features/services/ServicesScreen";
-export default ServicesScreen;
