+++ app/(private)/dashboard.tsx
@@ -1,0 +1,2 @@
+import DashboardScreen from "../../src/features/DashboardScreen/DashboardScreen";
+export default DashboardScreen;
