+++ app/(private)/investments.tsx
