+++ app/(private)/meus-cartoes.tsx
@@ -1,0 +1,3 @@
+import CardsScreen from "@/src/features/CardsScreen/CardsScreen";
+
+export default CardsScreen;
