+++ app/(private)/outros-servicos.tsx
@@ -1,0 +1,3 @@
+import ServicesScreen from "@/src/services/ServicesScreen";
+
+export default ServicesScreen;
