+++ app/_layout.tsx
@@ -1,5 +1,4 @@
 import { Slot } from "expo-router";
-import 'react-native-gesture-handler';
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { GlobalContextProvider } from "../src/contexts";
