+++ eslint.config.js
@@ -1,4 +1,4 @@
-// https://docs.expo.dev/guides/using-eslint/
+// eslint.config.js
 const { defineConfig } = require('eslint/config');
 const expoConfig = require('eslint-config-expo/flat');
 
@@ -6,5 +6,18 @@ module.exports = defineConfig([
   expoConfig,
   {
     ignores: ['dist/*'],
+    settings: {
+      'import/resolver': {
+        typescript: {
+          project: './tsconfig.json',
+        },
+        node: {
+          extensions: ['.js', '.jsx', '.ts', '.tsx', '.d.ts', '.svg'],
+        },
+      },
+    },
+    rules: {
+      'import/no-unresolved': 'error',
+    },
   },
 ]);
