+++ package.json
@@ -66,6 +66,8 @@
     "@types/react": "~19.0.10",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
+    "eslint-import-resolver-node": "^0.3.9",
+    "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-testing-library": "^7.6.3",
     "jest": "~29.7.0",
     "jest-expo": "~53.0.10",
