+++ src/features/home/HomeScreen.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { StyleSheet, View } from "react-native";
 import { tokens } from "../../theme/tokens";
-import { LoginForm } from "../auth/LoginForm";
+import { LoginForm } from "../auth/LoginForm/LoginForm";
 
 const styles = StyleSheet.create({
   container: {
