+++ src/shared/Footer/Footer.tsx
